requirements-dev.txt
# 开发依赖

# 代码质量
black==23.12.0
flake8==7.0.0
isort==5.13.2
mypy==1.7.1
pylint==3.0.3

# 测试增强
pytest-cov==4.1.0
pytest-rerunfailures==13.0
pytest-timeout==4.0.0
pytest-xdist==3.5.0

# 文档
mkdocs==1.5.3
mkdocs-material==9.5.3

# 性能分析
py-spy==0.3.14
memory-profiler==0.61.0

# 调试
ipdb==0.13.13
